nyc-taxi-app/requirements.txt
# This file defines the image's dependencies
Flask
google-cloud-pubsub




